email_producer/requirements.txt
langchain
kafka-python
openai
